Adding logging interceptor + connecting to Hiveserver not 2 some problem with the connection establishment

+++ src/main/java/com/spike/mapreduce/CustomerMapper.java
@@ -1,5 +1,6 @@
 package com.spike.mapreduce;
 
+import org.apache.hadoop.io.FloatWritable;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
@@ -12,10 +13,10 @@ import java.util.StringTokenizer;
  * Author: Amith Nambiar<amith.nmbr@gmail.com>
  * Date: 2/5/14
  */
-public class CustomerMapper extends MapReduceBase implements Mapper<LongWritable, Text, Text, IntWritable> {
+public class CustomerMapper extends MapReduceBase implements Mapper<LongWritable, Text, Text, FloatWritable> {
 
     @Override
-    public void map(LongWritable lineNumber, Text line, OutputCollector<Text, IntWritable> output, Reporter reporter) throws IOException {
+    public void map(LongWritable lineNumber, Text line, OutputCollector<Text, FloatWritable> output, Reporter reporter) throws IOException {
         String lineStr = line.toString();
         String strippedStr = lineStr.replaceAll("\\s", "");
         StringTokenizer tokenizer = new StringTokenizer(strippedStr, ",");
@@ -24,6 +25,6 @@ public class CustomerMapper extends MapReduceBase implements Mapper<LongWritable
         String customerName = tokenizer.nextToken();
         String amount = tokenizer.nextToken();
 
-        output.collect(new Text(customerName), new IntWritable(Integer.valueOf(amount)));
+        output.collect(new Text(customerName), new FloatWritable(Float.valueOf(amount)));
     }
 }
